[Update] 2-2 서비스 리스트 total_count 수정

+++ src/main/java/com/editdining/response/ListResult.java
@@ -8,6 +8,6 @@ import java.util.List;
 @Getter
 @Setter
 public class ListResult<T> extends CommonResult {
-    private int total_count;
+    private long total_count;
     private List<T> list;
 }
